requirements.txt
aiohttp
aiohttp_cors


# asynchroniczne podłączenie bazy danych
asyncpg

# asynchroniczny dostęp do sqlite
aiosqlite

# praca z obrazkami
Pillow

# asynchroniczny dostęp do bazy danych redis
aioredis

# biblioteka do performance-testów
locust